+++ src/com/intellij/idea/plugin/hybris/project/wizard/SelectHybrisModulesToImportStep.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -36,7 +36,7 @@ class SelectHybrisModulesToImportStep(wizard: WizardContext) : AbstractSelectMod
     private var selectionMode = ModuleDescriptorImportStatus.MANDATORY
 
     override fun init() {
-        fileChooser.addElementsMarkListener(ElementsChooser.ElementsMarkListener<ModuleDescriptor> { element, isMarked ->
+        fileChooser.addElementsMarkListener(ElementsChooser.ElementsMarkListener { element, isMarked ->
             if (element is YModuleDescriptor) {
                 if (isMarked) {
                     element.getAllDependencies()